[ms-abi] Leading VFPtrs don't suppress the leading zero sized flag
The MS-ABI tracks a bit that asserts that the first sub-object is zero
sized.  This bit is used to add padding between objects if there's the
potential for zero sized objects to alias.  The bit is still true even
if the zero sized base is lead by a VFPtr.  This patch makes clang mimic
that behavior.