When debugserver fails to attach to a process on a Darwin system, make a couple of additional checks to see if the attach was denied via the System Integrity Protection that is new in Mac OS X 10.11.  If so, return a special E87 error code to indicate this to lldb.
Up in lldb, if we receive the E87 error code, be specific
about why the attach failed.

Also detect the more common case of general attach failure
and print a better error message than "lost connection".

I believe this code will all build on Mac OS X 10.10 systems.
It may not compile or run on earlier versions of the OS.
None of this should build on other non-darwin systems.